backend/ml/requirements.txt
numpy>=1.26,<3
pandas>=2.1,<3
scikit-learn>=1.3,<2
joblib>=1.3,<2
openpyxl>=3.1,<4
Pillow>=10,<12
tensorflow>=2.13,<2.16
